version: v2.2.2版本发布 【新增】AI助理导入包 【优化】数字人小程序主页UI及操作逻辑 【优化】OEM站点LOGO同步 【优化】OEM站点标题自定 【优化】AI智能体提示词字数量开放2万字 【优化】RPA任务执行兼容性 【优化】微信朋友圈读取加载 【优化】微信朋友圈自动评论逻辑 【优化】AI美工图片返回占位样式 【优化】AI自动剪辑背景音乐处理 【优化】AI自动剪辑新增背景音乐包 【优化】RPA端进程中断问题

+++ server/app/api/logic/service/TokenLogService.php
@@ -18,11 +18,11 @@ class TokenLogService
     /**
      * 获取任务需要的算力
      * @param string $scene
-     * @return int
+     * @return float
      * @author L
      * @data 2024/7/31 16:34
      */
-    public static function getTypeScore(string $scene = ""): int
+    public static function getTypeScore(string $scene = ""): float
     {
         return ModelConfig::where('scene', $scene)->value('score', 0);
     }
@@ -31,12 +31,12 @@ class TokenLogService
      * @desc 检查用户token是否足够
      * @param int $uid
      * @param string $scene
-     * @return int
+     * @return float
      * @date 2024/7/29 16:15
      * @throws \Exception
      * @author dagouzi
      */
-    public static function checkToken(int $uid, string $scene = ""): int
+    public static function checkToken(int $uid, string $scene = ""): float
     {
         $use_token   = self::getTypeScore($scene);
         $userInfo = User::findOrEmpty($uid)->toArray();
